\name{BaSTA-package}
\alias{BaSTA-package}
\alias{BaSTA}

\docType{package}

\title{
 BaSTA: Parametric Bayesian estimation of age-specific survival for truncated and censored capture-recapture or census data.
}

\description{
 This package estimates age-specific mortality and survival trajectories with covariates from capture-recapture/recovery or census data under a Bayesian framework when many individuals are of unknown age. It includes tools for data checking, model diagnostics, and outputs such as life-tables and plots.
}

\details{
\tabular{ll}{
	Package: \tab BaSTA \cr
	Type: \tab Package\cr
	Version: \tab 2.0.3\cr
	Date: \tab 2025-06-27\cr
	License: \tab GNU General Public Licence\cr
	LazyLoad: \tab yes\cr
}

Parametric estimation of age-specific survival and mortality from capture-mark-recapture or census data including individuals with unknown times of birth and death.
}

\author{Fernando Colchero \email{fernando_colchero@eva.mpg.de}}

\references{
Colchero, F. and J.S. Clark (2012) Bayesian inference on age-specific survival from capture-recapture data for censored and truncated data. \emph{Journal of Animal Ecology} 81, 139-149.

Colchero, F., O.R Jones and M. Rebke. (2012) BaSTA: an R package for Bayesian estimation of age-specific survival from incomplete mark-recapture/recovery data with covariates. \emph{Methods in Ecology and Evolution} 3, 466-470.

Colchero, F., \emph{et al.}  (2021) The long lives of primates and the "invariant rate of aging" hypothesis. \emph{Nature Communications} 12:3666
}